refactor(web): split DesignShowcase into per-section files
Split apps/web/src/core/DesignShowcase.tsx (1245 LOC) into a small
orchestrator (70 LOC) plus 12 focused section files under
apps/web/src/core/designShowcase/. Largest section is now Feedback at
164 LOC (well under the 600-LOC threshold tracked in
docs/tech-debt/frontend.md §4).

Layout:
- DesignShowcase.tsx               — sticky header + section list
- designShowcase/_shared.tsx       — Sec / Group / ColorSwatch helpers
                                     and NAV_SECTIONS const
- designShowcase/sections/*.tsx    — one file per section (12)

Behaviour-preserving:
- Section state moves into the owning section component (Navigation
  owns tabVal/segVal/navActive; Overlays owns modal/sheet/confirm;
  Feedback owns switchOn; Celebration consumes useCelebration).
- The single external import path (apps/web/src/core/App.tsx, lazy
  import of './DesignShowcase' with named export DesignShowcase) is
  unchanged.
- Build, lint, and typecheck of the touched files all pass.
